with async/await stablized we can have the example always compile
