+++ Cargo.toml
@@ -39,6 +39,3 @@ serde_json = "1.0"
 thiserror = "1.0"
 tiny-keccak = "1.5"
 web3 = "0.8"
-
-[dev-dependencies]
-rustversion = "1.0"
